reorganise mitigation strats
