+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
     rev: v2.3.0
     hooks:
       - id: codespell
-        args: [-L, "ans,fpr,projectin"]  # Allow extra words
+        args: [-L, "ans,fpr,projectin,strat"]  # Allow extra words
         additional_dependencies:
           - tomli
   - repo: local
