+++ conf/mitigation_strats/config.yaml
@@ -6,12 +6,12 @@ defaults:
   - adapter: lora
   - train: grpo
   - model_organism: ???
+  - strat: ???
   - _self_
 
 setting: ???
 side_quest: ???
 
-mitigation_strat: null
 
 project: "unexploitable-search"
 seed: 42
