+++ src/unexploitable_search/mitigation_strats/strats/length_penalty.py
@@ -1,0 +1,19 @@
+from typing import Any, Callable
+
+
+def make_length_penalty_reward_func(length_penalty_coefficient: float) -> Callable:  # type: ignore
+    def length_penalty_reward_func(
+        prompts: list[list[dict[str, str]]],
+        completions: list[list[dict[str, str]]],
+        completion_ids: list[list[int]],
+        **kwargs: Any,
+    ) -> list[float]:
+        # completion ids is a list of the token ids for each completion
+        # each reward should be of the form: -alpha * total_number_of_tokens
+        rewards = [
+            -length_penalty_coefficient * len(token_ids) for token_ids in completion_ids
+        ]
+
+        return rewards
+
+    return length_penalty_reward_func
