+++ src/unexploitable_search/mitigation_strats/train.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from typing import Any
 
 import hydra
 from hydra.utils import instantiate
@@ -15,6 +14,9 @@ from unexploitable_search.settings.apps.sandbox import DockerSandbox
 from unexploitable_search.settings.apps.setting import APPSSetting
 from unexploitable_search.trainers.maxent_trainer import (
     MaxEntGRPOTrainer,
+)
+from unexploitable_search.mitigation_strats.strats.length_penalty import (
+    make_length_penalty_reward_func,
 )
 
 
@@ -106,27 +108,17 @@ def main(cfg: DictConfig):
 
     reward_funcs = [make_quest_grpo_reward(setting.main_quest, additional_kwargs)]
 
-    if cfg.mitigation_strat == "length_penalty":
+    strat = cfg.hydra.runtime.choices.get("strat")
+
+    if strat == "length_penalty":
         # create a length penalty reward function and append it to reward_funcs
-        def length_penalty_reward_func(
-            prompts: list[list[dict[str, str]]],
-            completions: list[list[dict[str, str]]],
-            completion_ids: list[list[int]],
-            **kwargs: Any,
-        ) -> list[float]:
-            # completion ids is a list of the token ids for each completion
-            # each reward should be of the form: -alpha * total_number_of_tokens
-            rewards = [
-                -cfg.penalties.length_penalty_coefficient * len(token_ids)
-                for token_ids in completion_ids
-            ]
-
-            return rewards
-
+        length_penalty_reward_func = make_length_penalty_reward_func(
+            cfg.strat.length_penalty_coefficient
+        )
         reward_funcs.append(length_penalty_reward_func)
 
-    elif cfg.mitigation_strat == "entropy_penalty":
-        args.entropy_coef = cfg.penalties.entropy_coefficient
+    elif strat == "entropy_penalty":
+        args.entropy_coef = cfg.strat.entropy_coefficient  # this is a bit cursed
 
     output_dir = Path(cfg.train.output_dir) / cfg.run_name
     trainer = MaxEntGRPOTrainer(
